Move {down,up}load code to {down,up}load.py
Also moves the DownloadableResource definition to model.py
